refactor(cardinal): move txpool out of types
